requirements.txt
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-queryfields==1.0.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-dotenv==0.18.0
